+++ src/core/models/comment.ts
@@ -1,0 +1,6 @@
+import { CommentRange } from 'typescript';
+
+export interface Comment {
+    text: string;
+    range: CommentRange;
+}