+++ src/core/models/import-node.ts
@@ -2,9 +2,15 @@ import {
     ImportDeclaration,
     LineAndCharacter
 } from 'typescript';
+import { Comment } from './comment';
 
 export interface ImportNode {
     importDeclaration: ImportDeclaration;
+    importComment: {
+        leadingComments: Comment[];
+        trailingComments: Comment[];
+    };
     start: LineAndCharacter;
     end: LineAndCharacter;
 }
+