+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+# [2.0.0]
+### Breaking changes
+* default sorting rules have been changed to:
+```json
+[
+    {
+        "type": "importMember",
+        "regex": "^$",
+        "orderLevel": 10,
+        "disableSort": true
+    },
+    {
+        "regex": "^[@]",
+        "orderLevel": 30
+    },
+    {
+        "regex": "^[.]",
+        "orderLevel": 40
+    }
+]
+```
+
+### New features
+* added support of the comments blocks.
+* added `importSorter.sortConfiguration.customOrderingRules.disableDefaultOrderSort` and `importSorter.sortConfiguration.customOrderingRules.rules.disableSort` which allows to
+disable sort for the group of imports
+### Bug fixes
+* fixing semi-relative import paths
+* correcting `newLineEachExpressionAfterCountLimit`
+
 # [1.2.0]
 ### New features
 * added `importSorter.generalConfiguration.exclude` to define the type of characters for new-lined imports